+++ LiftDataManager/LiftDataManager.csproj
@@ -60,12 +60,15 @@
     <Content Remove="Assets\WindowIcon.ico" />
   </ItemGroup>
   <ItemGroup>
+    <None Remove="Assets\Fluent\adminMode.png" />
+    <None Remove="Assets\Fluent\appDescription.png" />
     <None Remove="Assets\Fluent\controler.png" />
     <None Remove="Assets\Fluent\home.png" />
     <None Remove="Assets\Fluent\liftcar.png" />
     <None Remove="Assets\Fluent\liftdoors.png" />
     <None Remove="Assets\Fluent\shaft.png" />
     <None Remove="Assets\Fluent\text.png" />
+    <None Remove="Assets\Fluent\themeColor.png" />
     <None Remove="Assets\NavViewMenu\NavigationViewControlData.json" />
     <None Remove="Controls\CarDesignControl.xaml" />
     <None Remove="Controls\CarEquipmentControl.xaml" />
@@ -128,7 +131,7 @@
     <None Remove="Views\NutzlastberechnungPage.xaml" />
     <None Remove="Views\Settings\AboutSettingPage.xaml" />
     <None Remove="Views\Settings\GeneralSettingPage.xaml" />
-    <None Remove="Views\Settings\MaintenancePage.xaml" />
+    <None Remove="Views\Settings\MaintenanceSettingPage.xaml" />
     <None Remove="Views\Settings\ThemeSettingPage.xaml" />
   </ItemGroup>
   <ItemGroup>
@@ -204,11 +207,11 @@
     <PackageReference Include="CommunityToolkit.WinUI.Media" Version="8.0.240109" />
     <PackageReference Include="CommunityToolkit.WinUI.Triggers" Version="8.0.240109" />
     <PackageReference Include="CommunityToolkit.WinUI.UI.Controls.DataGrid" Version="7.1.2" />
-    <PackageReference Include="HtmlAgilityPack" Version="1.11.60" />
+    <PackageReference Include="HtmlAgilityPack" Version="1.11.61" />
     <PackageReference Include="Humanizer.Core" Version="2.14.1" />
     <PackageReference Include="Humanizer.Core.de" Version="2.14.1" />
     <PackageReference Include="Microsoft.Extensions.Hosting" Version="8.0.0" />
-    <PackageReference Include="Microsoft.WindowsAppSDK" Version="1.5.240404000" />
+    <PackageReference Include="Microsoft.WindowsAppSDK" Version="1.5.240428000" />
     <PackageReference Include="Microsoft.Xaml.Behaviors.WinUI.Managed" Version="2.0.9" />
     <PackageReference Include="Refractored.MvvmHelpers" Version="1.6.2" />
     <PackageReference Include="Serilog" Version="3.1.1" />
@@ -216,8 +219,8 @@
     <PackageReference Include="Serilog.Formatting.Compact" Version="2.0.0" />
     <PackageReference Include="Serilog.Sinks.Debug" Version="2.0.0" />
     <PackageReference Include="Serilog.Sinks.File" Version="5.0.0" />
-    <PackageReference Include="WinUICommunity.Components" Version="6.7.0" />
-    <PackageReference Include="WinUICommunity.Core" Version="6.7.0" />
+    <PackageReference Include="WinUICommunity.Components" Version="6.8.0" />
+    <PackageReference Include="WinUICommunity.Core" Version="6.8.0" />
     <PackageReference Include="SkiaSharp.Views.WinUI" Version="2.88.8" />
     <PackageReference Include="WinUIEx" Version="2.3.4" />
   </ItemGroup>
@@ -306,7 +309,7 @@
       <XamlRuntime>$(DefaultXamlRuntime)</XamlRuntime>
       <SubType>Designer</SubType>
     </Page>
-    <Page Update="Views\Settings\MaintenancePage.xaml">
+    <Page Update="Views\Settings\MaintenanceSettingPage.xaml">
       <XamlRuntime>$(DefaultXamlRuntime)</XamlRuntime>
       <SubType>Designer</SubType>
     </Page>
