--- LiftDataManager/ViewModels/Settings/MaintenanceViewModel.cs
@@ -1,8 +1,0 @@
-﻿namespace LiftDataManager.ViewModels;
-
-public partial class MaintenanceViewModel : ObservableRecipient
-{
-    public MaintenanceViewModel()
-    {
-    }
-}
