+++ LiftDataManager/Views/Settings/AboutSettingPage.xaml
@@ -2,11 +2,63 @@
     x:Class="LiftDataManager.Views.AboutSettingPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:controls="using:CommunityToolkit.WinUI.Controls"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:models="using:LiftDataManager.Core.DataAccessLayer.Models"
+    xmlns:sys="using:System"
+    xmlns:wuc="using:WinUICommunity"
     mc:Ignorable="d">
-
+    <Page.Resources>
+        <DataTemplate x:Key="VersionsHistoryItemTemplate" x:DataType="models:LiftDataManagerVersion">
+            <controls:SettingsCard>
+                <controls:SettingsCard.HeaderIcon>
+                    <SymbolIcon Symbol="RepeatOne" />
+                </controls:SettingsCard.HeaderIcon>
+                <controls:SettingsCard.Header>
+                    <StackPanel Margin="0,0,0,5" Orientation="Horizontal">
+                        <TextBlock Style="{StaticResource ListTitleStyle}" Text="{x:Bind VersionsNumber}" />
+                        <TextBlock Margin="20,0,0,0" Text="{x:Bind sys:String.Format('{0:d}', VersionsDate)}" />
+                    </StackPanel>
+                </controls:SettingsCard.Header>
+                <controls:SettingsCard.Description>
+                    <TextBlock Margin="70,0,0,0" Text="{x:Bind VersionDescription}" />
+                </controls:SettingsCard.Description>
+                <StackPanel Margin="0,0,20,-20" VerticalAlignment="Bottom">
+                    <TextBlock
+                        FontSize="12"
+                        FontWeight="SemiBold"
+                        Text="Author :" />
+                    <TextBlock FontSize="12" Text="{x:Bind Author}" />
+                </StackPanel>
+            </controls:SettingsCard>
+        </DataTemplate>
+    </Page.Resources>
     <Grid x:Name="ContentArea">
-        <TextBlock FontSize="36" Text="About" />
+        <ScrollViewer>
+            <StackPanel Margin="0,0,15,0">
+                <controls:SettingsCard Header="{x:Bind ViewModel.VersionDescription, Mode=OneWay}" HeaderIcon="{wuc:BitmapIcon Source=Assets/Fluent/appDescription.png}">
+                    <controls:SettingsCard.Description>
+                        <StackPanel>
+                            <TextBlock Margin="0,10,0,0" Style="{ThemeResource BodyTextBlockStyle}">
+                                Anwendung zur Datenerfassung und Bearbeitung der Aufzugsspezifikationsdaten.<LineBreak />
+                                Datenselektion, Plausibilitätsprüfung und Berechnungsmodule sorgen für eine optimale Datenintegrität.</TextBlock>
+                            <TextBlock
+                                Margin="0,10,0,0"
+                                Style="{ThemeResource BodyTextBlockStyle}"
+                                Text="©Buchholz Roland" />
+                        </StackPanel>
+                    </controls:SettingsCard.Description>
+                </controls:SettingsCard>
+                <Expander
+                    Padding="0,5,0,0"
+                    HorizontalAlignment="Stretch"
+                    HorizontalContentAlignment="Stretch"
+                    Header="Versions History"
+                    IsExpanded="True">
+                    <ListView ItemTemplate="{StaticResource VersionsHistoryItemTemplate}" ItemsSource="{x:Bind ViewModel.VersionsHistory, Mode=OneWay}" />
+                </Expander>
+            </StackPanel>
+        </ScrollViewer>
     </Grid>
 </Page>
