--- LiftDataManager/Views/Settings/MaintenancePage.xaml
@@ -1,12 +1,0 @@
-﻿<Page
-    x:Class="LiftDataManager.Views.MaintenancePage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d">
-
-    <Grid x:Name="ContentArea">
-        <TextBlock FontSize="36" Text="MaintenancePage" />
-    </Grid>
-</Page>
