+++ LiftDataManager/Views/Settings/MaintenanceSettingPage.xaml.cs
@@ -1,0 +1,15 @@
+﻿namespace LiftDataManager.Views;
+
+public sealed partial class MaintenanceSettingPage : Page
+{
+    public MaintenanceSettingViewModel ViewModel
+    {
+        get;
+    }
+
+    public MaintenanceSettingPage()
+    {
+        ViewModel = App.GetService<MaintenanceSettingViewModel>();
+        InitializeComponent();
+    }
+}
