+++ LiftDataManager/Views/Settings/ThemeSettingPage.xaml
@@ -2,11 +2,82 @@
     x:Class="LiftDataManager.Views.ThemeSettingPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:Core="using:Microsoft.Xaml.Interactions.Core"
+    xmlns:Interactivity="using:Microsoft.Xaml.Interactivity"
+    xmlns:controls="using:CommunityToolkit.WinUI.Controls"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:wuc="using:WinUICommunity"
     mc:Ignorable="d">
 
     <Grid x:Name="ContentArea">
-        <TextBlock FontSize="36" Text="ThemeSettingPage" />
+        <ScrollViewer>
+            <StackPanel Margin="0,0,15,0">
+                <StackPanel Margin="{StaticResource SmallTopBottomMargin}">
+                    <controls:SettingsCard
+                        Description="Wählen Sie das Thema, das Ihrer Stimmung und Ihren Vorlieben entspricht. Sie können zwischen dunklen, hellen oder Systemthemen wählen."
+                        Header="App Theme"
+                        HeaderIcon="{wuc:BitmapIcon Source=Assets/Fluent/theme.png}">
+                        <ComboBox x:Name="CmbTheme">
+                            <ComboBoxItem Content="Light" Tag="Light" />
+                            <ComboBoxItem Content="Dark" Tag="Dark" />
+                            <ComboBoxItem Content="Default" Tag="Default" />
+                            <Interactivity:Interaction.Behaviors>
+                                <Core:EventTriggerBehavior EventName="SelectionChanged">
+                                    <Core:EventTriggerBehavior.Actions>
+                                        <Core:InvokeCommandAction Command="{x:Bind ViewModel.ThemeChangedCommand}" CommandParameter="{Binding ElementName=CmbTheme}" />
+                                    </Core:EventTriggerBehavior.Actions>
+                                </Core:EventTriggerBehavior>
+                            </Interactivity:Interaction.Behaviors>
+                        </ComboBox>
+                    </controls:SettingsCard>
+                    <controls:SettingsCard
+                        Description="Ändern Sie das Erscheinungsbild des Hintergrundmaterials hinter Ihrer App. Sie können zwischen Glimmer-, Acryl- oder transparenten Optionen wählen."
+                        Header="App Material"
+                        HeaderIcon="{wuc:BitmapIcon Source=Assets/Fluent/backdrop.png}">
+                        <ComboBox x:Name="CmbBackdrop">
+                            <ComboBoxItem Content="None" Tag="None" />
+                            <ComboBoxItem Content="Mica" Tag="Mica" />
+                            <ComboBoxItem Content="Mica Alt" Tag="MicaAlt" />
+                            <ComboBoxItem Content="Desktop Acrylic" Tag="DesktopAcrylic" />
+                            <ComboBoxItem Content="Acrylic Base" Tag="AcrylicBase" />
+                            <ComboBoxItem Content="Acrylic Thin" Tag="AcrylicThin" />
+                            <ComboBoxItem Content="Transparent" Tag="Transparent" />
+                            <Interactivity:Interaction.Behaviors>
+                                <Core:EventTriggerBehavior EventName="SelectionChanged">
+                                    <Core:EventTriggerBehavior.Actions>
+                                        <Core:InvokeCommandAction Command="{x:Bind ViewModel.BackdropChangedCommand}" CommandParameter="{Binding ElementName=CmbBackdrop}" />
+                                    </Core:EventTriggerBehavior.Actions>
+                                </Core:EventTriggerBehavior>
+                            </Interactivity:Interaction.Behaviors>
+                        </ComboBox>
+                    </controls:SettingsCard>
+                    <controls:SettingsCard
+                        Description="Nuzten Sie die Windows-Akzentfarbe oder die App-Standardfarbe."
+                        Header="App Akzentfarbe"
+                        HeaderIcon="{wuc:BitmapIcon Source=Assets/Fluent/themeColor.png}">
+                        <ToggleSwitch
+                            x:Name="SwitchAccentColor"
+                            Margin="0,0,20,0"
+                            IsOn="{x:Bind ViewModel.CustomAccentColor, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
+                            OffContent="Standardfarbe Blau"
+                            OnContent="Windows-Akzentfarbe">
+                            <Interactivity:Interaction.Behaviors>
+                                <Core:EventTriggerBehavior EventName="Toggled">
+                                    <Core:InvokeCommandAction Command="{x:Bind ViewModel.SwitchAccentColorCommand}" />
+                                </Core:EventTriggerBehavior>
+                            </Interactivity:Interaction.Behaviors>
+                        </ToggleSwitch>
+                    </controls:SettingsCard>
+                    <controls:SettingsCard
+                        ActionIcon="{wuc:BitmapIcon Source=Assets/Fluent/external.png}"
+                        Command="{x:Bind ViewModel.OpenWindowsColorSettingsCommand}"
+                        Description="Ändern Sie die Farbe Ihrer Windows-Akzentfarbe. Sie können die Windows-Akzentfarbeinstellung öffnen, indem Sie auf diese Karte klicken."
+                        Header="System Akzentfarbe"
+                        HeaderIcon="{wuc:BitmapIcon Source=Assets/Fluent/color.png}"
+                        IsClickEnabled="True" />
+                </StackPanel>
+            </StackPanel>
+        </ScrollViewer>
     </Grid>
 </Page>
