+++ LiftDataManager/Views/Settings/ThemeSettingPage.xaml.cs
@@ -11,5 +11,12 @@ public sealed partial class ThemeSettingPage : Page
     {
         ViewModel = App.GetService<ThemeSettingViewModel>();
         InitializeComponent();
+        Loaded += ThemeSettingPage_Loaded;
+    }
+
+    private void ThemeSettingPage_Loaded(object sender, RoutedEventArgs e) 
+    {
+        ViewModel.SetDefaultTheme(CmbTheme);
+        ViewModel.SetDefaultBackdrop(CmbBackdrop);
     }
 }
