+++ LiftDataManager/Views/SettingsPage.xaml.cs
@@ -12,12 +12,5 @@ public sealed partial class SettingsPage : Page
     {
         ViewModel = App.GetService<SettingsViewModel>();
         InitializeComponent();
-        Loaded += SettingsPage_Loaded;
-    }
-
-    private void SettingsPage_Loaded(object sender, RoutedEventArgs e)
-    {
-        ViewModel._themeService.SetThemeComboBoxDefaultItem(CmbTheme);
-        ViewModel._themeService.SetBackdropComboBoxDefaultItem(CmbBackdrop);
     }
 }
