+++ LiftDataManager/Views/UserControls/BreadcrumbBarUserControl.xaml
@@ -1,0 +1,27 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<UserControl
+    x:Class="LiftDataManager.Views.BreadcrumbBarUserControl"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:i="using:Microsoft.Xaml.Interactivity"
+    xmlns:ic="using:Microsoft.Xaml.Interactions.Core"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <BreadcrumbBar Margin="10" ItemsSource="{x:Bind ViewModel.BreadcrumbBarCollection, Mode=OneWay}">
+        <BreadcrumbBar.ItemTemplate>
+            <DataTemplate x:DataType="x:String">
+                <TextBlock Style="{ThemeResource TitleTextBlockStyle}" Text="{x:Bind}" />
+            </DataTemplate>
+        </BreadcrumbBar.ItemTemplate>
+        <i:Interaction.Behaviors>
+            <ic:EventTriggerBehavior EventName="ItemClicked">
+                <ic:EventTriggerBehavior.Actions>
+                    <ic:InvokeCommandAction Command="{x:Bind ViewModel.ItemClickCommand}" />
+                </ic:EventTriggerBehavior.Actions>
+            </ic:EventTriggerBehavior>
+        </i:Interaction.Behaviors>
+    </BreadcrumbBar>
+</UserControl>
+
